分类下还有商品，不能删除

+++ jeepay-service/src/main/java/com/jeequan/jeepay/service/impl/ProductCategoryService.java
@@ -2,8 +2,10 @@ package com.jeequan.jeepay.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.update.LambdaUpdateWrapper;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+import com.jeequan.jeepay.core.constants.CS;
 import com.jeequan.jeepay.core.entity.Product;
 import com.jeequan.jeepay.core.entity.ProductCategory;
+import com.jeequan.jeepay.core.exception.BizException;
 import com.jeequan.jeepay.service.mapper.ProductCategoryMapper;
 import com.jeequan.jeepay.service.mapper.ProductMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,6 +29,11 @@ public class ProductCategoryService extends ServiceImpl<ProductCategoryMapper, P
 
     @Transactional
     public void removeCategory(Long categoryId) {
+        Integer count = productMapper.selectCount(Product.gw().eq(Product::getCategoryId, categoryId).eq(Product::getDeleteState, CS.NO));
+        if (count > 0) {
+            throw new BizException("该分类下还有商品，不能删除");
+        }
+
         // 删除分类
         productCategoryMapper.deleteById(categoryId);
         // 移除分类下的商品
